use portable integer types (cstdint) instead of LONG
LONG -> int32_t

LONG is another Windows.h typedef. It is 32 bits and signed on both Windows
targets, so int32_t is an exact replacement.

One member is spelled plain long instead: CStatesManager::nRef, the reference
count of the Direct3D effect state manager. It is passed to
InterlockedIncrement and InterlockedDecrement, which take long*, and int32_t
is int on MSVC, so an int32_t* argument would not bind.

+++ Versions/Temporary/Engine/Sources/ED_B2_M1/HeightWindowV3.h
@@ -54,7 +54,7 @@ class CHeightWindowV3 : public CResizeDialog, public ICommandHandler
 	afx_msg void OnUpdateHeights();
 	afx_msg void OnItemchangedTileList( NMHDR* pNMHDR, LRESULT* pResult );
 
-	LONG nHeightTimer;
+	int32_t nHeightTimer;
 	uint32_t dwHeightData;
   inline UINT GetHeightID() { return 2; }
   inline UINT GetHeightTimerInterval() { return 100; }	// Частота в миллисекундах
